// frontend/src/core/api/generated/model/shipmentRowStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Constructor ERP API
 * OpenAPI spec version: 1.0
 */

export type ShipmentRowStatus = (typeof ShipmentRowStatus)[keyof typeof ShipmentRowStatus]

export const ShipmentRowStatus = {
  RECORDED: 'RECORDED',
  CANCELLED: 'CANCELLED',
} as const